Do not save objects to file beyond a certain file size Invocation count file saves count across separate workload executions

+++ pankti-instrument/src/main/java/se/kth/castor/pankti/instrument/plugins/MethodAspect0.java
@@ -4,28 +4,38 @@ import org.glowroot.agent.plugin.api.*;
 import org.glowroot.agent.plugin.api.weaving.*;
 
 import java.io.*;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.Scanner;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class MethodAspect0 {
     private static int INVOCATION_COUNT;
+    private static boolean fileSizeWithinLimits = true;
 
     @Pointcut(className = "fully.qualified.path.to.class",
             methodName = "methodToInstrument",
             methodParameterTypes = {"param1", "param2"},
             timerName = "Timer - name")
-    public static class PureMethodAdvice implements AdviceTemplate {
-        private static final TimerName timer = Agent.getTimerName(PureMethodAdvice.class);
-        private static final String transactionType = "Pure";
+    public static class TargetMethodAdvice implements AdviceTemplate {
+        private static final TimerName timer = Agent.getTimerName(TargetMethodAdvice.class);
+        private static final String transactionType = "Target";
         private static final int COUNT = 0;
         private static String receivingObjectFilePath;
         private static String paramObjectsFilePath;
         private static String returnedObjectFilePath;
         private static String invocationCountFilePath;
         private static String invokedMethodsCSVFilePath;
-        private static Logger logger = Logger.getLogger(PureMethodAdvice.class);
+        private static Logger logger = Logger.getLogger(TargetMethodAdvice.class);
         private static String rowInCSVFile = "";
-        private static final String methodFQN = PureMethodAdvice.class.getAnnotation(Pointcut.class).className() + "."
-                + PureMethodAdvice.class.getAnnotation(Pointcut.class).methodName();
+        private static final String methodParamTypesString = String.join(",", TargetMethodAdvice.class.getAnnotation(Pointcut.class).methodParameterTypes());
+        private static final String methodFQN = TargetMethodAdvice.class.getAnnotation(Pointcut.class).className() + "."
+                + TargetMethodAdvice.class.getAnnotation(Pointcut.class).methodName();
+        private static final String invocationString = String.format("Invocation count for %s: ", methodFQN);
+
+        // TODO: add params to filenames
 
         private static void setup() {
             AdviceTemplate.setUpXStream();
@@ -35,6 +45,15 @@ public class MethodAspect0 {
             returnedObjectFilePath = fileNames[2];
             invocationCountFilePath = fileNames[3];
             invokedMethodsCSVFilePath = fileNames[4];
+            checkFileSizeLimit(receivingObjectFilePath);
+        }
+
+        // Limit object XML files to ~400 MB
+        public static void checkFileSizeLimit(String fileName) {
+            File file = new File(fileName);
+            if (file.exists() && (file.length() / (1024 * 1024) > 350)) {
+                fileSizeWithinLimits = false;
+            }
         }
 
         public static synchronized void writeObjectXMLToFile(Object objectToWrite, String objectFilePath) {
@@ -57,9 +76,22 @@ public class MethodAspect0 {
 
         public static synchronized void writeInvocationCountToFile() {
             try {
-                FileWriter objectFileWriter = new FileWriter(invocationCountFilePath);
-                objectFileWriter.write("Invocation count for " + methodFQN + " : " + INVOCATION_COUNT);
-                objectFileWriter.close();
+                File invocationCountFile = new File(invocationCountFilePath);
+
+                if (!invocationCountFile.exists()) {
+                    FileWriter objectFileWriter = new FileWriter(invocationCountFilePath);
+                    objectFileWriter.write(invocationString + INVOCATION_COUNT);
+                    objectFileWriter.close();
+                } else {
+                    String content = Files.readString(Paths.get(invocationCountFilePath));
+                    Pattern pattern = Pattern.compile(invocationString + "(\\d+$)");
+                    Matcher matcher = pattern.matcher(content);
+                    if (matcher.find()) {
+                        int count = Integer.parseInt(matcher.group(1)) + 1;
+                        content = matcher.replaceAll(invocationString + count);
+                        Files.write(invocationCountFile.toPath(), content.getBytes(StandardCharsets.UTF_8));
+                    }
+                }
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -93,14 +125,14 @@ public class MethodAspect0 {
                                           @BindReceiver Object receivingObject,
                                           @BindParameterArray Object parameterObjects,
                                           @BindMethodName String methodName) {
-            if (INVOCATION_COUNT < 50) {
-                setup();
+            setup();
+            if (fileSizeWithinLimits) {
                 writeObjectXMLToFile(receivingObject, receivingObjectFilePath);
                 writeObjectXMLToFile(parameterObjects, paramObjectsFilePath);
             }
             MessageSupplier messageSupplier = MessageSupplier.create(
                     "className: {}, methodName: {}",
-                    PureMethodAdvice.class.getAnnotation(Pointcut.class).className(),
+                    TargetMethodAdvice.class.getAnnotation(Pointcut.class).className(),
                     methodName
             );
             return context.startTransaction(transactionType, methodName, messageSupplier, timer, OptionalThreadContext.AlreadyInTransactionBehavior.CAPTURE_NEW_TRANSACTION);
@@ -109,8 +141,9 @@ public class MethodAspect0 {
         @OnReturn
         public static void onReturn(@BindReturn Object returnedObject,
                                     @BindTraveler TraceEntry traceEntry) {
-            if (INVOCATION_COUNT < 50) {
+            if (fileSizeWithinLimits) {
                 writeObjectXMLToFile(returnedObject, returnedObjectFilePath);
+                checkFileSizeLimit(receivingObjectFilePath);
             }
             INVOCATION_COUNT++;
             if (INVOCATION_COUNT == 1) {
